Add the newest websockets and multiboard states

See merge request cse1105/2022-2023/teams/oopp-team-12!44

+++ server/src/main/java/server/services/BoardService.java
@@ -2,6 +2,7 @@ package server.services;
 
 import commons.Board;
 import commons.CardList;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import server.api.BoardController;
 import server.database.BoardRepository;
@@ -11,9 +12,12 @@ import java.util.logging.Logger;
 
 @Service
 public class BoardService {
+    private static final Logger log = Logger.getLogger(BoardController.class.getName());
 
     private BoardRepository boardRepository;
-    static Logger log = Logger.getLogger(BoardController.class.getName());
+
+    @Autowired
+    SynchronizationService synchronizationService;
 
     public BoardService(BoardRepository boardRepository) {
         this.boardRepository = boardRepository;
@@ -27,24 +31,27 @@ public class BoardService {
         if (id < 0 || boardRepository.findById(id).isEmpty()) {
             throw new IllegalArgumentException("Invalid board id:" + id);
         }
-        return boardRepository.findById(id).orElse(null);
+        var board = boardRepository.findById(id).get();
+        return board;
     }
 
     public Board saveBoard(Board board) {
         if (board == null) {
             throw new IllegalArgumentException("Board cannot be null");
         }
+
+        synchronizationService.addBoardToUpdate(board.getId());
         return boardRepository.save(board);
     }
 
-    public CardList saveCardList(CardList list, long boardId){
+    public CardList saveCardList(CardList list, long boardId) {
         if (list == null) {
             throw new IllegalArgumentException("Trying to add null list");
         }
 
         try {
             var boardOptional = this.getBoardById(boardId);
-        } catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("Trying to add list into non-existent board");
         }
 
@@ -60,16 +67,20 @@ public class BoardService {
     }
 
     public void deleteBoardById(long id) {
-        if (id < 0 || boardRepository.findById(id).orElse(null) == null){
+        if (id < 0 || boardRepository.findById(id).orElse(null) == null) {
             throw new IllegalArgumentException("Invalid board id:" + id);
         }
+
+        if (id == 1)
+            throw new IllegalArgumentException("Deleting default board is prohibited");
+
         boardRepository.deleteById(id);
     }
 
-    public void deleteCardListById(long listId, long boardId){
-        try{
+    public void deleteCardListById(long listId, long boardId) {
+        try {
             var boardOptional = this.getBoardById(boardId);
-        } catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("Trying to delete list from non-existent board");
         }
 
@@ -86,10 +97,10 @@ public class BoardService {
         throw new IllegalArgumentException("Trying to delete non-existent list");
     }
 
-    public void moveCard (long cardId, long listId, int index, long boardId){
-        try{
+    public void moveCard(long cardId, long listId, int index, long boardId) {
+        try {
             var boardOptional = this.getBoardById(boardId);
-        } catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("Trying to move card in non-existent board");
         }
 
@@ -132,27 +143,27 @@ public class BoardService {
         board.sync();
         sourceList.sync();
         targetList.sync();
-        boardRepository.save(board);
+        saveBoard(board);
     }
 
-    public Board update(Board board, long id){
+    public Board update(Board board, long id) {
         if (board == null) {
             throw new IllegalArgumentException("Board is null");
         }
         log.info("Updating board: " + board.getId());
         try {
             var boardOptional = this.getBoardById(id);
-        } catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("Trying to update non existing board");
         }
 
-        if (id != board.getId()){
+        if (id != board.getId()) {
             throw new IllegalArgumentException("Ids are not coherent in board update");
         }
 
         try {
             var stored = this.getBoardById(id);
-        } catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("Something went wrong while updating board");
         }
